+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - New asset allocation report.
 - New --cost flag for transactions.
 - New portfolio performance report.
+- New portfolio summary report.
 
 ### Changed